Fix playbook calls to ardana-start and ardana-stop (bsc#1095455)
In HOS5.0 and prior, there was odd logic in the backend
of the ardana-service that converted underscores to dashes
(and in some cases vice versa) to "normalize" playbook calls.
That logic doesn't exist in the refactored version
of the ardana-service. The playbook calls to ardana-start
and ardana-stop were incorrect in the UI previously but
were "fixed" in the backend via the dash-underscore_
logic... This commit fixes the frontend now that
the workaround logic is gone from the backend.

Change-Id: I8baf1cda382e63f741becb2e459c283a84adab61