Do not overwrite activation status based on completed plays (bsc#1095599)
The ops console is using completed plays to determine the state
of a host, even if those plays happened some time ago, and
is prioritizing deactivate over activate. The problem is
that if the UI is ever used to deactivate a host, it will
be marked as deactivated forevermore.

To work around this for now, only use active playbooks
to override the backend status of a host. This will retain
the ability for someone reloading the UI to see that
a host is in the process of activating or deactivating,
while deferring the status of hosts that are not presently
changing state to the normal means of getting status.

This does have the side effect of an activated or deactivated
host being marked incorrectly for about two minutes after
such a process completes, as the backend takes some time
to finish updating.

Change-Id: Ibca97ba3d063ce4dcbbfd6a6c3e03bf5b6559574