Fix deprecated references to 'alive' (bsc#1102789)
The ardana-service no longer returns the 'alive' field from the play
APIs.  This field can be inferred by the presence of either the 'code'
field or the 'endTime' field, both of which are present if and only if
the process is no longer alive. The opsconsole UI was never updated to
reflect this change.

Partial-Bug: bsc#1102789
Change-Id: Id28e609f72d33364a2e1066d86b1f043242b7929